backend/services/datscan_service.py
"""
DATScan Analysis Service
Analyzes dopamine transporter brain scans for Parkinson's disease detection
"""

import logging
import numpy as np
from typing import Dict, Any, Optional
import tempfile
import os
from PIL import Image
import io
import base64

logger = logging.getLogger(__name__)

class DATScanAnalysisService:
    """Service for analyzing DATScan images using deep learning"""
    
    def __init__(self):
        self.model = None
        self.feature_names = None
        self._load_model()
    
    def _load_model(self):
        """Load the pre-trained DATScan model"""
        try:
            # For now, we'll use a mock model
            # In production, this would load a real deep learning model
            logger.info("✅ DATScan model loaded (mock implementation)")
            self.model = "mock_datscan_model"
            self.feature_names = ["datscan_dl_features"]
        except Exception as e:
            logger.error(f"Failed to load DATScan model: {e}")
            self.model = None
    
    async def analyze_datscan(self, image_path: str) -> Dict[str, Any]:
        """Analyze DATScan image for Parkinson's disease detection"""
        try:
            if not self.model:
                raise Exception("DATScan model not loaded")
            
            # Load and preprocess image
            image = self._load_image(image_path)
            features = self._extract_features(image)
            
            # Mock prediction (replace with actual model inference)
            prediction_result = self._mock_prediction(features)
            
            # Add metadata
            result = {
                **prediction_result,
                "model_type": "DATScan Analysis",
                "processing_time": 2.5,  # Mock processing time
                "clinical_notes": self._generate_clinical_notes(prediction_result),
                "key_features": [
                    {
                        "feature_name": "dopamine_activity",
                        "importance_score": 0.95,
                        "description": "Dopamine transporter binding in striatum",
                        "category": "neurotransmitter"
                    },
                    {
                        "feature_name": "striatal_asymmetry",
                        "importance_score": 0.87,
                        "description": "Left-right asymmetry in striatal uptake",
                        "category": "spatial"
                    },
                    {
                        "feature_name": "background_ratio",
                        "importance_score": 0.82,
                        "description": "Striatal to background uptake ratio",
                        "category": "quantitative"
                    }
                ]
            }
            
            logger.info("✅ DATScan analysis completed successfully")
            return result
            
        except Exception as e:
            logger.error(f"DATScan analysis failed: {e}")
            raise
    
    def _load_image(self, image_path: str) -> Image.Image:
        """Load and validate image"""
        try:
            image = Image.open(image_path)
            
            # Convert to RGB if necessary
            if image.mode != 'RGB':
                image = image.convert('RGB')
            
            # Resize to standard dimensions (224x224 for typical DL models)
            image = image.resize((224, 224), Image.Resampling.LANCZOS)
            
            return image
            
        except Exception as e:
            raise Exception(f"Failed to load image: {e}")
    
    def _extract_features(self, image: Image.Image) -> np.ndarray:
        """Extract features from image (mock implementation)"""
        # Convert image to numpy array
        img_array = np.array(image)
        
        # Mock feature extraction
        # In production, this would use a real deep learning model
        features = np.random.rand(512)  # Mock 512-dimensional feature vector
        
        return features
    
    def _mock_prediction(self, features: np.ndarray) -> Dict[str, Any]:
        """Mock prediction (replace with actual model inference)"""
        # Simulate model prediction
        np.random.seed(hash(str(features.tobytes())) % 2**32)
        
        # Generate realistic-looking probabilities
        base_probability = 0.3  # Base probability for PD
        feature_influence = np.mean(features) * 0.4  # Feature influence
        
        probability_pd = np.clip(base_probability + feature_influence, 0.1, 0.9)
        probability_healthy = 1.0 - probability_pd
        
        # Determine prediction
        prediction = 1 if probability_pd > 0.5 else 0
        prediction_label = "Parkinson's Disease" if prediction == 1 else "Healthy"
        
        # Calculate confidence based on feature consistency
        confidence = 0.6 + (np.std(features) * 0.3)  # Higher std = lower confidence
        confidence = np.clip(confidence, 0.3, 0.95)
        
        return {
            "prediction": prediction,
            "prediction_label": prediction_label,
            "confidence": confidence,
            "probability_pd": probability_pd,
            "probability_healthy": probability_healthy
        }
    
    def _generate_clinical_notes(self, prediction_result: Dict[str, Any]) -> str:
        """Generate clinical interpretation notes"""
        prediction = prediction_result['prediction']
        confidence = prediction_result['confidence']
        probability_pd = prediction_result['probability_pd']
        
        if prediction == 1:  # Parkinson's Disease
            if confidence > 0.8:
                severity = "high confidence"
            elif confidence > 0.6:
                severity = "moderate confidence"
            else:
                severity = "low confidence"
            
            notes = f"DATScan analysis shows {severity} for reduced dopamine transporter activity "
            notes += f"in the striatum (probability: {probability_pd:.1%}). "
            
            if probability_pd > 0.7:
                notes += "Significant reduction in striatal uptake consistent with PD diagnosis. "
            elif probability_pd > 0.5:
                notes += "Moderate reduction in striatal uptake suggestive of early PD. "
            else:
                notes += "Mild reduction in striatal uptake requiring clinical correlation. "
                
        else:  # Healthy
            if confidence > 0.8:
                notes = "DATScan analysis shows normal dopamine transporter activity "
                notes += f"in the striatum (confidence: {confidence:.1%}). "
                notes += "No significant abnormalities detected."
            else:
                notes = "DATScan analysis suggests normal dopamine transporter activity, "
                notes += "though with limited confidence. Clinical correlation recommended."
        
        return notes
