backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
librosa==0.10.1
xgboost>=1.5.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
joblib>=1.1.0
pydantic>=2.0.0
pydub>=0.25.0
soundfile>=0.12.0
python-dotenv>=0.19.0
aiofiles>=0.8.0
Pillow>=9.0.0
opencv-python>=4.5.0
